Estilos arreglados, navbar arreglado y pagina para cotizar

+++ database/migrations/2023_08_22_194119_create_precio_unitarios_table.php
@@ -18,7 +18,7 @@ return new class () extends Migration {
             ->onDelete('cascade');
             $table->float('medida');
             $table->integer('cantidad_colores');
-            $table->float('precio');
+            $table->decimal('precio');
             $table->timestamp('fecha_desde')->useCurrent();
             $table->timestamp('fecha_hasta')->default(null)->nullable()->index();
         });
